Split Name::Spelling tests out of name_test.rb (#4708)
Moves the 6 tests exercising Name::Spelling (parent_if_parent_deprecated,
names_from_synonymous_genera, suggest_alternate_spellings, guess_word/
guess_with_errors via test_name_guessing and test_guess_name_with_colon_in_pattern,
fix_self_referential_misspellings) into test/models/name/spelling_test.rb,
matching app/models/name/spelling.rb. Includes a local copy of
create_test_name (used by test_suggest_alternate_spelling), same
pattern as the Taxonomy split.

require_relative'd from name_test.rb so running that file alone still
runs the full original suite.
